20231005 湖南万汇 直接同步中台商品数据到ERP逻辑完成

+++ src/main/java/hykx/ds/whkc/entity/MyGoodsEntity.java
@@ -20,7 +20,7 @@ public class MyGoodsEntity {
     private int pch;
     private String yxq;
     private int sl;
-    private int lsj;
+    private float lsj;
     private float dj;
     private String dw;
     private String jx;
